Move filenames out of `compiler/main.ml`
The previous version of `compiler/main.ml` had embedded file names which
makes the use of the program less flexible (as well as more difficult to
understand). This new version requires the user to pass the input and
output filepaths at the command line when executing the binary. For
example, if this is run in the `compiler/` directory, it will compile
the `a.bc` file and put the output in the same place as before:

    $ ./main \
        ../test_bench/programs/code/code/a.bc \
	../test_bench/programs/code/asm/a.bc.s

Note that it would be (imo) a bit more convenient if this used the
pretty standard unix `-o` flag, but I have no idea how to do that in
ocaml.